02_Building_a_Multimodal_AI_Model_CLIP/requirements.txt
torch
Pillow
transformers
